History Implementation B, non-breaking, with alternate component
