+++ packages/react/src/index.ts
@@ -2,3 +2,4 @@ export * from "./context";
 export * from "./icons";
 export * from "./receipt-blocks";
 export * from "./receipt";
+export * from "./receipt-history";
