+++ packages/react/src/receipt-history/index.ts
@@ -1,0 +1,1 @@
+export * from "./receipt-with-history";
